refactor: create a manager for a submit event
create a manager for a submit event

MODULE-1-PART-3-gallery
